Module cut::path

Enums

Error 🔒

Functions

deep_copy 🔒

Recursively copy the contents of src to dst. Fails if src transitively contains a symlink. Only copies paths that pass the keep predicate.

normalize_path 🔒

Normalize the representation of path by eliminating redundant . components and applying .. components. Does not access the filesystem (e.g. to resolve symlinks or test for file existence), unlike std::fs::canonicalize.

path_relative_to 🔒

Return the path to dst relative to src. If src is a file, the path is relative to the directory that contains it, while if it is a directory, the path is relative to it. Returns an error if either src or dst do not exist.

shortest_new_prefix 🔒

Returns the shortest prefix of path that doesn’t exist, or None if path already exists.