Struct cut::plan::CutPlan

pub(crate) struct CutPlan {
    root: PathBuf,
    directories: BTreeSet<PathBuf>,
    packages: BTreeMap<String, CutPackage>,
}

Description of where packages should be copied to, what their new names should be, and whether they should be added to the workspace members or exclude fields.

Fields
root: PathBuf

Root of the repository, where the Cargo.toml containing the workspace configuration is found.

directories: BTreeSet<PathBuf>

New directories that need to be created. Used to clean-up copied packages on roll-back. If multiple nested directories must be created, only contains their shortest common prefix.

packages: BTreeMap<String, CutPackage>

Mapping from the names of existing packages to be cut, to the details of where they will be copied to.

Implementations

impl CutPlan

pub(crate) fn discover(args: Args) -> Result<Self>

Scan args.directories looking for args.packages to produce a new plan. The resulting plan is guaranteed not to contain any duplicate packages (by name or path), or overwrite any existing packages. Returns an error if it’s not possible to construct such a plan.

pub(crate) fn execute(&self) -> Result<()>

Copy the packages according to this plan. On success, all the packages will be copied to their destinations, and their dependencies will be fixed up. On failure, pending changes are rolled back.

fn execute_(&self) -> Result<()>

fn copy_package(&self, package: &CutPackage) -> Result<()>

Copy the contents of package from its src_path to its dst_path, unchanged.

fn update_package(&self, package: &CutPackage) -> Result<()>

Fix the contents of the copied package’s Cargo.toml: name altered to match package.dst_name and local relative-path-based dependencies are updated to account for the copied package’s new location. Assumes that all copied files exist (but may not contain up-to-date information).

fn update_dependencies( &self, src_path: impl AsRef<Path>, dst_path: impl AsRef<Path>, table: &mut dyn TableLike ) -> Result<()>

Find all dependency tables in table, part of a manifest at dst_path/Cargo.toml (originally at src_path/Cargo.toml), and fix (relative) paths to account for the change in the package’s location.

fn update_dependency( &self, src_path: impl AsRef<Path>, dst_path: impl AsRef<Path>, dep_name: KeyMut<'_>, dep: &mut Item ) -> Result<()>

Update an individual dependency from a copied package manifest. Only local path-based dependencies are updated:

Dep = { path = "..." }

If Dep is another package to be copied as part of this plan, the path is updated to the location it is copied to. Otherwise, its location (a relative path) is updated to account for the fact that the copied package is at a new location.

fn update_workspace(&self) -> Result<()>

Add entries to the members and exclude arrays in the root manifest’s workspace table.

fn rollback(&self)

Attempt to clean-up the partial results of executing a plan, by deleting the directories that the plan would have created. Swallows and prints errors to make sure as much clean-up as possible is done – this function is typically called when some other error has occurred, so it’s unclear what it’s starting state would be.