cut::args
Struct Args

pub(crate) struct Args {
    pub feature: String,
    pub root: Option<PathBuf>,
    pub directories: Vec<Directory>,
    pub packages: Vec<String>,
    pub workspace_update: bool,
    pub dry_run: bool,
}

Tool for cutting duplicate versions of a subset of crates in a git repository.

Duplicated crate dependencies are redirected so that if a crate is duplicated with its dependency, the duplicate’s dependency points to the duplicated dependency. Package names are updated to avoid conflicts with their original. Duplicates respect membership or exclusion from a workspace.

Fields
feature: String

Name of the feature the crates are being cut for – duplicated crate package names will be suffixed with a hyphen followed by this feature name.

root: Option<PathBuf>

Root of repository – all source and destination paths must be within this path, and it must contain the repo’s workspace configuration. Defaults to the parent of the working directory that contains a .git directory.

directories: Vec<Directory>

Add a directory to duplicate crates from, along with the destination to duplicate it to, and optionally a suffix to remove from package names within this directory, all separated by colons.

Only crates (directories containing a Cargo.toml file) found under the source (first) path whose package names were supplied as a --package will be duplicated at the destination (second) path. Copying will preserve the directory structure from the source directory to the destination directory.

packages: Vec<String>

Package names to include in the cut (this must match the package name in its source location, including any suffixes)

workspace_update: bool

Don’t make changes to the workspace.

dry_run: bool

Don’t execute the cut, just display it.