cut
Module plan

Structs

CutPackage 🔒

Details for an individual copied package in the feature being cut.

CutPlan 🔒

Description of where packages should be copied to, what their new names should be, and whether they should be added to the workspace members or exclude fields.

Workspace 🔒

Relevant contents of a Cargo.toml workspace section.

Enums

Error 🔒

WorkspaceState 🔒

Whether the package in question is an explicit member of the workspace, an explicit exclude of the workspace, or neither (in which case it could still transitively be one or the other).

Functions

discover_root 🔒

Find the root of the git repository containing cwd, if it exists, return None otherwise. This function only searches prefixes of the provided path for the git repo, so if the path is given as a relative path within the repository, the root will not be found.

format_array_of_strings 🔒

Format a TOML array of strings: Splits elements over multiple lines, indents them, sorts them, and adds a trailing comma.

package_name 🔒

path_to_toml_value 🔒

Represent path as a TOML value, by first describing it as a relative path (relative to root), and then converting it to a String. Fails if either root or path are not real paths (cannot be canonicalized), or the resulting relative path cannot be represented as a String.

toml_path_array_to_set 🔒

Read [field] from table, as an array of strings, and interpret as a set of paths, canonicalized relative to a root path.